# trading-cards/api/src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"AccountNftResponse": {
				"properties": {
					"nft": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object"
			},
			"SuccessfulResult_AccountNftResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"result": {
						"$ref": "#/components/schemas/AccountNftResponse"
					}
				},
				"required": [
					"success",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_FailedResult.Exclude_keyofFailedResult.errorMessage__": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"errorCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_FailedResult.errorMessage_": {
				"$ref": "#/components/schemas/Pick_FailedResult.Exclude_keyofFailedResult.errorMessage__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"ApiResult_AccountNftResponse_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessfulResult_AccountNftResponse_"
					},
					{
						"$ref": "#/components/schemas/Omit_FailedResult.errorMessage_"
					}
				]
			},
			"IGetOwnedCardsResult": {
				"description": "'GetOwnedCards' return type",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"owner_nft_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"registry_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"owner_nft_id",
					"registry_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetCardsResponse": {
				"properties": {
					"cards": {
						"items": {
							"$ref": "#/components/schemas/IGetOwnedCardsResult"
						},
						"type": "array"
					}
				},
				"required": [
					"cards"
				],
				"type": "object"
			},
			"SuccessfulResult_GetCardsResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"result": {
						"$ref": "#/components/schemas/GetCardsResponse"
					}
				},
				"required": [
					"success",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResult_GetCardsResponse_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessfulResult_GetCardsResponse_"
					},
					{
						"$ref": "#/components/schemas/Omit_FailedResult.errorMessage_"
					}
				]
			},
			"numberArray": {
				"items": {
					"type": "number",
					"format": "double"
				},
				"type": "array"
			},
			"IGetBoughtPacksResult": {
				"description": "'GetBoughtPacks' return type",
				"properties": {
					"buyer_nft_id": {
						"type": "number",
						"format": "double"
					},
					"card_registry_ids": {
						"$ref": "#/components/schemas/numberArray"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"buyer_nft_id",
					"card_registry_ids",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetPacksResponse": {
				"properties": {
					"packs": {
						"items": {
							"$ref": "#/components/schemas/IGetBoughtPacksResult"
						},
						"type": "array"
					}
				},
				"required": [
					"packs"
				],
				"type": "object"
			},
			"SuccessfulResult_GetPacksResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"result": {
						"$ref": "#/components/schemas/GetPacksResponse"
					}
				},
				"required": [
					"success",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResult_GetPacksResponse_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessfulResult_GetPacksResponse_"
					},
					{
						"$ref": "#/components/schemas/Omit_FailedResult.errorMessage_"
					}
				]
			},
			"IGetTradeNftsResult": {
				"description": "'GetTradeNfts' return type",
				"properties": {
					"cards": {
						"allOf": [
							{
								"$ref": "#/components/schemas/numberArray"
							}
						],
						"nullable": true
					},
					"nft_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"cards",
					"nft_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGetCardsByIdsResult": {
				"description": "'GetCardsByIds' return type",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"owner_nft_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"registry_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"owner_nft_id",
					"registry_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.IGetCardsByIdsResult_": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/IGetCardsByIdsResult"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"GetTradeNftsResponse": {
				"properties": {
					"cardLookup": {
						"$ref": "#/components/schemas/Record_string.IGetCardsByIdsResult_"
					},
					"tradeNfts": {
						"items": {
							"$ref": "#/components/schemas/IGetTradeNftsResult"
						},
						"type": "array"
					}
				},
				"required": [
					"cardLookup",
					"tradeNfts"
				],
				"type": "object"
			},
			"SuccessfulResult_GetTradeNftsResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"result": {
						"$ref": "#/components/schemas/GetTradeNftsResponse"
					}
				},
				"required": [
					"success",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResult_GetTradeNftsResponse_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessfulResult_GetTradeNftsResponse_"
					},
					{
						"$ref": "#/components/schemas/Omit_FailedResult.errorMessage_"
					}
				]
			},
			"IGetUserStatsResult": {
				"description": "'GetUserStats' return type",
				"properties": {
					"losses": {
						"type": "number",
						"format": "double"
					},
					"nft_id": {
						"type": "number",
						"format": "double"
					},
					"ties": {
						"type": "number",
						"format": "double"
					},
					"wins": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"losses",
					"nft_id",
					"ties",
					"wins"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserStats": {
				"$ref": "#/components/schemas/IGetUserStatsResult"
			},
			"UserStatsResponse": {
				"properties": {
					"stats": {
						"$ref": "#/components/schemas/UserStats"
					}
				},
				"required": [
					"stats"
				],
				"type": "object"
			},
			"SuccessfulResult_UserStatsResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"result": {
						"$ref": "#/components/schemas/UserStatsResponse"
					}
				},
				"required": [
					"success",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResult_UserStatsResponse_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessfulResult_UserStatsResponse_"
					},
					{
						"$ref": "#/components/schemas/Omit_FailedResult.errorMessage_"
					}
				]
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@cards/api",
		"version": "1.0.0",
		"description": "Webserver REST API for the middleware to interact with game node.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user/accountNft": {
			"get": {
				"operationId": "GetWalletNFT",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResult_AccountNftResponse_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "wallet",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/cards": {
			"get": {
				"operationId": "GetCards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResult_GetCardsResponse_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "nftId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/packs": {
			"get": {
				"operationId": "GetPacks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResult_GetPacksResponse_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "nftId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/tradeNfts": {
			"get": {
				"operationId": "GetTradeNfts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResult_GetTradeNftsResponse_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "nftId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/stats": {
			"get": {
				"operationId": "Stats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResult_UserStatsResponse_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "nftId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}